Rename the python package name to easydiffusion (from sd_internal)
