+++ ui/easydiffusion/utils/__init__.py
@@ -1,0 +1,8 @@
+import logging
+
+log = logging.getLogger('easydiffusion')
+
+from .save_utils import (
+    save_images_to_disk,
+    get_printable_request,
+)
